requirements.txt
Flask==3.0.3
requests==2.31.0
Faker==25.4.0
python-dotenv==1.0.1
# these are the dependecies for the above main libraries
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.2
